H.264 Encode:  Enable profile and level supported
OpenMAX component for H.264 encoder should enable a index used by
Stagefright for encoding.

The index OMX_IndexParamVideoProfileLevelQuerySupported should
respond with all the profiles and levels supported by the encoder.
So far only BP profile and level 1 are enabled.

Modification in omxil_core is to expose the functionality required
to know the profile/level currently supported by the H.264 port

Change-Id: Iddfe0a77fd0c466e501d348806a9cf5a869bcbb7