omxil-core: Load shared libraries upfront
Sandboxing on Chromium OS requires that libraries are loaded
at the creation of objects before it is locked down.

The libraries to be used by the plug-in hierarchy are made
available and are passed down to the parser.

Due to the locking of sandbox, instead of having the configuration
file to add components, the components are loaded from a table
defined within the header file.

Change-Id: I0488677a854a9c14a94af0b922ef3d48e11c99c3
Signed-off-by: Daniel Charles <daniel.charles@intel.com>