Fix hang when port disable called for empty port.
In gst-omx playback, when transitioning component to idle state
the output port is disabled. The call to disable output port
hangs in WaitPortBufferCompletion() in the OMX IL Core. This
is due to a bug where WaitPortBufferCompletion() waits for
buffer free operation to complete on an empty port. Since port
is empty buffer free will never be called. Added check on buffer
count before doing the wait to avoid the hang.

Change-Id: Id572250e4a57c23de28a4bf02fc7f88f3b94a484
Signed-off-by: Sameer Kibey <sameer.kibey@intel.com>