componentbase: cease port after port setting changed
this change follows openamx il specification 1.2 chapter 3.1.1.4.4
we need cease decoder when we report a "port settings changed"
this give il client a chance to reallocate buffers and reconfigure component.
image a case: we report  a port setting changed, il client start to
reconfigure the component,then we find and report a decode error.
It will force il client handle error in the middle of reconfigureation.
For gst-omx, it just crashed in this case.